Move 'task' package out of FastAPI and update imports
