+++ examples/fastlib/resources/base.py
@@ -2,8 +2,8 @@ from typing import NoReturn
 
 from cuenca_validations.errors import WrongCredsError
 
+from agave.core.exc import UnauthorizedError
 from agave.fastapi import RestApiBlueprint
-from agave.fastapi.exc import UnauthorizedError
 
 app = RestApiBlueprint()
